+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Added a data dictionary and data export task for admins
 - **General Changes**
   - Improved display of uploaded user, highlight, partner, and team member photos
+- **Design Changes**
+  - Moving to fixed-width version of site
 
 ## 8.3.0 (October 13, 2015)