+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 8.4.0
+## 8.4.0 (October 19, 2015)
 
 ### Enhancements
 - **Admin Changes**